decouple instantsearch logic from autocomplete component
