+++ React InstantSearch Hooks/query-suggestions/src/App.tsx
@@ -1,5 +1,3 @@
-import algoliasearch from 'algoliasearch/lite';
-
 import {
   Configure,
   CurrentRefinements,
@@ -8,23 +6,20 @@ import {
   InstantSearch,
 } from 'react-instantsearch-hooks-web';
 
-import { Autocomplete } from './Autocomplete';
-
 import './App.css';
 
 import type { Hit } from 'instantsearch.js';
 
-const searchClient = algoliasearch(
-  'latency',
-  '6be0576ff61c053d5f9a3225e2a90f76'
-);
+import { QuerySuggestions } from './QuerySuggestions';
+import { searchClient } from './searchClient';
 
 export function App() {
   return (
     <div className="container">
       <InstantSearch searchClient={searchClient} indexName="instant_search">
         <Configure hitsPerPage={16} />
-        <Autocomplete searchClient={searchClient} />
+
+        <QuerySuggestions />
 
         <CurrentRefinements
           className="current-refinements"
