+++ React InstantSearch Hooks/query-suggestions/src/searchClient.ts
@@ -1,0 +1,6 @@
+import algoliasearch from 'algoliasearch/lite';
+
+export const searchClient = algoliasearch(
+  'latency',
+  '6be0576ff61c053d5f9a3225e2a90f76'
+);
